[CMA-122] feat : 유저별 운동한 암장 리스트 조회 api 연결

+++ app/src/main/java/com/climus/climeet/presentation/ui/main/record/stats/SelectGymAdapter.kt
@@ -11,7 +11,7 @@ import com.climus.climeet.databinding.ItemStatsGymBinding
 import com.climus.climeet.presentation.ui.main.record.model.SelectGymData
 
 class SelectGymAdapter(
-    private val selectedItemId: Long,
+    private val selectedItemId: Int,
     private val itemClickListener: (SelectGymData) -> Unit
 ) :
     ListAdapter<SelectGymData, SelectGymAdapter.SelectGymViewHolder>(diffCallback) {
